requirements.txt
fastapi
uvicorn[standard]
pydantic
langchain
langchain-google-genai
python-dotenv